Public Documentation

Documentation for the submodule Brooglie.jl.

This module is a solver for the time-independent Schödinger equation and a copy of a project https://github.com/RedPointyJackson/Brooglie (MIT License).

Index

CarrierCapture.Brooglie

CarrierCapture.Brooglie.buildH

CarrierCapture.Brooglie.solve

Public Interface

CarrierCapture.Brooglie — Module

A copy of https://github.com/RedPointyJackson/Brooglie

CarrierCapture.Brooglie.buildH — Method

buildH(V; N=20, a=-1, b=1, m=1)

Hamiltonian of a particle of mass m in a box spanning from a to b in all D dimensions with a basis of N elements (number of partitions of the space in each coordinate). The potential V(x, y, z, ...) is a function of D arguments.

CarrierCapture.Brooglie.solve — Method

solve(V; N=500, a=-1, b=1, m=1, nev=N÷20, maxiter=1000)

Solve the potential V(x,y,z,...) in a grid xᵢ ∈ [a,b], discretized in N steps.

The particle is assumed to have mass m (by default 1, a electron mass).

The function will return the nev first energy levels (in Hartree[1]) and its normalized eigenfunctions.

[1] A Hartree is equivalent to 27.21… eV. The global variable H2eV, equal to that value, can be accessed under Brooglie.H2eV for convenience.