Public Documentation

Documentation for CarrierCapture.jl's public interface.

Contents

Public Documentation

Contents

Index

Public Interface

Index

CarrierCapture

CarrierCapture.conf_coord

CarrierCapture.potential

Public Interface

link to Brooglie.jl Documentation

CarrierCapture — Module

Main module for CarrierCapture.jl – A set of codes to compute carrier capture and recombination rates in semiconducting compounds.

Two structs are exported from this module for public use:

potential. Potential.

conf_coord. Configuration coordinate.

Exports

calc_capt_coeff!

calc_overlap!

cc_from_dict

conf_coord

double_well

find_crossing

fitHarmonicParams

fitMorseParams

fit_pot!

getHarmonicCapture

getHarmonicQ_m

getMorseCapture

getMorseQ_m

get_bspline

get_spline

harmonic

morse

plot_cc!

plot_ccs

plot_pot!

plot_pots

polyfunc

pot_from_dict

potential

solve1D_ev_amu

solve_pot!

sqwell

CarrierCapture.potential — Type

Stores a potential in one-dimensional space Q, with discreet points (E0, Q0) and fitting function func.

Fields

name – the name of potential.

color – the color for plotting.

QE_data – the (n X 2) DataFrame of data points (Q vs Energy).

func_type – the type of fitting function ("bspline", "spline", "harmonic", "polyfunc", "morse_poly", "morse").

params – the list of hyper paramters for the fitting function.

p0 – the initial paramters for the fitting function.

Q, E – Q and E=func(Q, p_opt; params).

neV – the number of eigenvalues to be evaluated.

ϵ – the list of eigenvalues

Constructor

potential()

CarrierCapture.conf_coord — Type

Stores two potentials with e-ph coupling constant W to calculate the capture coefficient capt_coeff(temperature).

Fields

name – the name of conf_coord.

V1 and V2 – the initial and fianal potentials.

W – the e-ph coupling matrix element.

g – the degeneracy.

temperature – the temperature range where capt_coeff is calculated.

capt_coeff – the capture coefficient.

Constructor

conf_coord(pot_i::potential, pot_f::potential)